requirements.txt
anyio==4.8.0
bleach==6.2.0
bokeh==3.6.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorcet==3.1.0
contourpy==1.3.1
cryptography==44.0.0
exceptiongroup==1.2.2
holoviews==1.20.0
hvplot==0.11.2
idna==3.10
Jinja2==3.1.5
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
panel==1.5.5
param==2.2.0
pillow==11.1.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.3.0
watchfiles==1.0.4
webencodings==0.5.1
xyzservices==2024.9.0
